# lib/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Jambonz REST API
  description: Jambonz REST API specification
  contact:
    email: daveh@drachtio.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: /v1
    description: jambonz API server
...
paths:
...
  /VoipCarriers:
    post:
      tags:
        - Voip Carriers
      summary: create voip carrier
      operationId: createVoipCarrier
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: voip carrier name
                  example: fastco
                description:
                  type: string
                  example: my US sip trunking provider
                account_sid:
                  type: string
                application_sid:
                  type: string
                e164_leading_plus:
                  type: boolean
                  description: whether a leading + is required on INVITEs to this provider
                  example: true
                requires_register:
                  type: boolean
                  description: wehther this provider requires us to send a REGISTER to them in order to receive calls
                register_use_tls:
                  type: boolean
                  description: wehther this provider requires us to send a REGISTER use TLS protocol
                register_username:
                  type: string
                  description: sip username to authenticate with, if registration is required
                  example: foo
                register_sip_realm:
                  type: string
                  description: sip realm to authenticate with, if registration is required
                  example: sip.fastco.com
                register_password:
                  type: string
                  description: sip password to authenticate with, if registration is required
                  example: bar
                register_from_user:
                  type: string
                  description: optional username to apply in From header
                register_from_domain:
                  type: string
                  description: optional domain to apply in From header
                register_public_ip_in_contact:
                  type: boolean
                  description: if true, use our public ip in Contact header; otherwise, use sip realm
                register_expires:
                  type: integer
                  description: optional SIP registration expiry interval in seconds; overrides the global default (3600) for this carrier
                tech_prefix:
                  type: string
                  description: prefix to be applied to the called number for outbound call attempts
                inbound_auth_username:
                  type: string
                  description: challenge inbound calls with this username/password if supplied
                inbound_auth_password:
                  type: string
                  description: challenge inbound calls with this username/password if supplied
                diversion:
                  type: string
                  description: Diversion header or phone number to apply to outbound calls
              required: 
                - name
      responses:
        201:
          description: voip carrier successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulAdd'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      tags:
        - Voip Carriers
      summary: list voip carriers
      operationId: listVoipCarriers
      responses:
        200:
          description: list of voip carriers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoipCarrier'
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /VoipCarriers/{VoipCarrierSid}:
    parameters:
      - name: VoipCarrierSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
    delete:
      tags:
        - Voip Carriers
      summary: delete a voip carrier
      operationId: deleteVoipCarrier
      responses:
        204:
          description: voip carrier successfully deleted
        404:
          description: voip carrier not found
        422:
          description: unprocessable entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
                example:
                  msg: a service provider with active accounts can not be deleted
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    get:
      tags:
        - Voip Carriers
      summary: retrieve voip carrier
      operationId: getVoipCarrier
      responses:
        200:
          description: voip carrier found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoipCarrier'
        404:
          description: voip carrier not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
    put:
      tags:
        - Voip Carriers
      summary: update voip carrier
      operationId: updateVoipCarrier
      parameters:
      - name: VoipCarrierSid
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoipCarrier'
      responses:
        204:
          description: voip carrier updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoipCarrier'
        400:
          description: bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
        404:
          description: voip carrier not found
        500:
          description: system error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralError'
  /VoipCarriers/{VoipCarrierSid}/pcap/by-carrier-sid:
    parameters:
      - name: VoipCarrierSid
        in: path
        required: true
        schema:
          type: string
    get:
      tags:
        - Voip Carriers
      summary: retrieve pcap for a registration-based carrier
      description: >
        Looks up the SIP Call-ID from the carrier's register_status field,
        then retrieves the REGISTER pcap from Homer.
      operationId: getCarrierPcapByCarrierSid
      responses:
        200:
          description: pcap file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        400:
          description: Homer not configured
        404:
          description: carrier not found, no registration status, or no pcap available
...
  /ServiceProviders/{ServiceProviderSid}/VoipCarriers:
    parameters:
      - name: ServiceProviderSid
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags:
        - Service Providers
      summary: get all carriers for a service provider
      operationId: getServiceProviderCarriers
      responses:
        200:
          description: account listing
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VoipCarrier'
        403:
          description: unauthorized
        404:
          description: service provider not found
    post:
      tags:
        - Service Providers
      summary: create a carrier
      operationId: createCarrierForServiceProvider
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoipCarrier'
      responses:
        201:
          description: service provider successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulAdd'
        404:
          description: service provider not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
  schemas:
...
    SuccessfulAdd:
      type: object
      required:
        - sid
      properties:
        sid:
          type: string
      example:
        sid: 9d26a637-1679-471f-8da8-7150266e1254
    GeneralError:
      type: object
      required:
        - msg
      properties:
        msg:
          type: string
      example:
        msg: specific error detail will be provided here
...
    VoipCarrier:
      type: object
      properties:
        voip_carrier_sid:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        account_sid:
          type: string
          format: uuid
        application_sid:
          type: string
          format: uuid          
        e164_leading_plus:
          type: boolean
        requires_register:
          type: boolean
        register_username:
          type: string
        register_sip_realm:
          type: string
        register_password:
          type: string
        tech_prefix:
          type: string
        inbound_auth_username:
          type: string
        inbound_auth_password:
          type: string
        diversion:
          type: string
        is_active:
          type: boolean
        smpp_system_id:
          type: string
        smpp_password:
          type: string
        smpp_inbound_system_id:
          type: string
        smpp_inbound_password:
          type: string
        smpp_enquire_link_interval:
          type: number
          format: integer
        trunk_type:
          type: string
          enum:
            - static_ip
            - auth
            - reg
          description: >
            Defines the trunk type: static_ip (IP-based, default), auth (digest authentication),
            or reg (carrier registers to us)
        register_from_user:
          type: string
        register_from_domain:
          type: string
        register_public_ip_in_contact:
          type: boolean
        register_expires:
          type: integer
          nullable: true
          description: SIP registration expiry interval in seconds; overrides the global default for this carrier
        outbound_sip_proxy:
          type: string
      required:
        - voip_carrier_sid
        - name
...
security:
  - bearerAuth: []